Added distance sensor

+++ src/I2CDevice.cpp
@@ -10,8 +10,8 @@ void I2CDevice::init() {
     this->handle = wiringPiI2CSetup(this->id);
 }
 
-ssize_t I2CDevice::send(const void *data, size_t size) {
-    return write(this->handle, data, size);
+int I2CDevice::send(const void *data, size_t size) {
+    return (int) write(this->handle, data, size);
 }
 
 int I2CDevice::writeReg8(int reg, int data) {
