+++ src/I2CDevice.h
@@ -10,7 +10,7 @@ public:
     I2CDevice(int id);
 
     virtual void init();
-    ssize_t send(const void *data, size_t size);
+    int send(const void *data, size_t size);
     int writeReg8(int reg, int data);
     int readReg8(int reg);
